webhook/requirements.txt
Flask==1.0.2
plists==0.0.4
request==1.0.2
requests==2.20.1
mysql-connector==2.1.6
mysql-connector-python==8.0.13
